fix auth bug, add root admin access control

+++ resources/views/admin-side/adminRole.blade.php
@@ -2,14 +2,16 @@
 
 @section('content')
 <h1 style="text-align: center">ADMIN ROLE</h1>
-<a class="btn btn-primary top-table-btn" href="{{route('admin-role.register')}}">Add New Admin</a>
-<button class="btn btn-danger top-table-btn" type="button" onclick="deleteAdminOnPress({{$admins->count()}})" id="delete-admin-btn">Delete Admin</button>
 @if (session('message'))
     <div class="alert alert-success">
         {{ session('message') }}
     </div>
 @endif
-<table class="table table-striped">
+<a class="btn btn-primary top-table-btn" href="{{route('admin-role.register')}}">Add New Admin</a>
+@if (Session::get('auth-email') == 'diahdewi.165@gmail.com')
+    <button class="btn btn-danger top-table-btn" type="button" onclick="deleteAdminOnPress({{$admins->count()}})" id="delete-admin-btn">Delete Admin</button>
+@endif
+<table class="table table-striped" id="admin-table">
   <thead style="background-color: #279636; color: white">
     <tr>
       <th scope="col" width="80%">Admin</th>
@@ -23,9 +25,13 @@
         <form method="post" action="{{route('admin-role.delete')}}">
         <td>{{$admin->email}}</td>
         <input type="hidden" value="{{$admin->email}}" name="email"/>
-        <td><button type="button" class="btn btn-danger table-btn" data-toggle="modal" data-target="#confirmationModal" id="delete-btn-no-{{$index}}" >Delete</button></td>
+        <td>
+        @if (Session::get('auth-email') == 'diahdewi.165@gmail.com')
+            <button type="button" class="btn btn-danger table-btn" data-toggle="modal" data-target="#confirmationModal_{{$index}}" id="delete-btn-no-{{$index}}" >Delete</button>
+        @endif
+        </td>
         <!-- Modal -->
-        <div class="modal fade" id="confirmationModal" role="dialog">
+        <div class="modal fade" id="confirmationModal_{{$index}}" role="dialog">
             <div class="modal-dialog">
             
             <!-- Modal content-->
@@ -35,7 +41,7 @@
                 </div>
 
                 <div class="modal-body">
-                    <p>Deleted account cannot be restored. Are you sure want to delete the account?</p>
+                    <p>Deleted account cannot be restored. Are you sure want to delete {{$admin->email}} admin account?</p>
                 </div>
 
                 <div class="modal-footer text-right">
